+++ packages_generated/container/src/v1beta1/marshalling.gen.ts
@@ -130,6 +130,8 @@ export const unmarshalContainer = (data: unknown): Container => {
   }
 
   return {
+    args: data.args,
+    command: data.command,
     cpuLimit: data.cpu_limit,
     createdAt: unmarshalDate(data.created_at),
     description: data.description,
@@ -150,6 +152,7 @@ export const unmarshalContainer = (data: unknown): Container => {
     namespaceId: data.namespace_id,
     port: data.port,
     privacy: data.privacy,
+    privateNetworkId: data.private_network_id,
     protocol: data.protocol,
     readyAt: unmarshalDate(data.ready_at),
     region: data.region,
@@ -229,6 +232,7 @@ export const unmarshalNamespace = (data: unknown): Namespace => {
     status: data.status,
     tags: data.tags,
     updatedAt: unmarshalDate(data.updated_at),
+    vpcIntegrationActivated: data.vpc_integration_activated,
   } as Namespace
 }
 
@@ -482,6 +486,8 @@ export const marshalCreateContainerRequest = (
   request: CreateContainerRequest,
   defaults: DefaultValues,
 ): Record<string, unknown> => ({
+  args: request.args,
+  command: request.command,
   cpu_limit: request.cpuLimit,
   description: request.description,
   environment_variables: request.environmentVariables,
@@ -499,6 +505,7 @@ export const marshalCreateContainerRequest = (
   namespace_id: request.namespaceId,
   port: request.port,
   privacy: request.privacy,
+  private_network_id: request.privateNetworkId,
   protocol: request.protocol,
   registry_image: request.registryImage,
   sandbox: request.sandbox,
@@ -538,6 +545,7 @@ export const marshalCreateNamespaceRequest = (
   request: CreateNamespaceRequest,
   defaults: DefaultValues,
 ): Record<string, unknown> => ({
+  activate_vpc_integration: request.activateVpcIntegration,
   description: request.description,
   environment_variables: request.environmentVariables,
   name: request.name || randomName('cns'),
@@ -637,6 +645,8 @@ export const marshalUpdateContainerRequest = (
   request: UpdateContainerRequest,
   defaults: DefaultValues,
 ): Record<string, unknown> => ({
+  args: request.args,
+  command: request.command,
   cpu_limit: request.cpuLimit,
   description: request.description,
   environment_variables: request.environmentVariables,
@@ -652,6 +662,7 @@ export const marshalUpdateContainerRequest = (
   min_scale: request.minScale,
   port: request.port,
   privacy: request.privacy,
+  private_network_id: request.privateNetworkId,
   protocol: request.protocol,
   redeploy: request.redeploy,
   registry_image: request.registryImage,