+++ packages_generated/dedibox/src/v1/types.gen.ts
@@ -814,6 +814,10 @@ export interface Server {
    * Whether or not the server is already part of an rpnv2 group.
    */
   isRpnv2Member: boolean
+  /**
+   * Whether or not the server is HDS.
+   */
+  isHds: boolean
 }
 
 export interface FailoverBlock {
@@ -1414,6 +1418,10 @@ export interface ServerSummary {
    * Supported RPN version.
    */
   rpnVersion?: number
+  /**
+   * Whether or not the server is HDS.
+   */
+  isHds: boolean
 }
 
 export interface CPU {