+++ packages_generated/dedibox/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-dedibox@1.1.0...@scaleway/sdk-dedibox@1.2.0) (2025-05-26)
+
+### Features
+
+- **dedibox:** add a is_hds flag for ddx2elts servers list ([#2154](https://github.com/scaleway/scaleway-sdk-js/issues/2154)) ([6c16e21](https://github.com/scaleway/scaleway-sdk-js/commit/6c16e210968479b811c85fcec41a81a9c9fe8beb))
+
+# [1.1.0](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-dedibox@1.0.3...@scaleway/sdk-dedibox@1.1.0) (2025-05-19)
+
+### Features
+
+- **dedibox:** add a is_hds flag for ddx2elts servers ([#2145](https://github.com/scaleway/scaleway-sdk-js/issues/2145)) ([21d538f](https://github.com/scaleway/scaleway-sdk-js/commit/21d538f790b794851eee1f71f49fbd8de70b9379))
+
+## [1.0.3](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-dedibox@1.0.2...@scaleway/sdk-dedibox@1.0.3) (2025-05-13)
+
+**Note:** Version bump only for package @scaleway/sdk-dedibox
+
 ## [1.0.2](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-dedibox@1.0.1...@scaleway/sdk-dedibox@1.0.2) (2025-05-13)
 
 ### Bug Fixes