+++ packages_generated/function/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.3.2 (2025-05-26)
+
+**Note:** Version bump only for package @scaleway/sdk-function
+
+## [1.3.1](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-function@1.3.0...@scaleway/sdk-function@1.3.1) (2025-05-13)
+
+**Note:** Version bump only for package @scaleway/sdk-function
+
 # [1.3.0](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-function@1.0.1...@scaleway/sdk-function@1.3.0) (2025-05-13)
 
 ### Bug Fixes