+++ packages_generated/key_manager/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-key-manager@1.2.0...@scaleway/sdk-key-manager@1.2.1) (2025-05-26)
+
+**Note:** Version bump only for package @scaleway/sdk-key-manager
+
+# [1.2.0](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-key-manager@1.1.1...@scaleway/sdk-key-manager@1.2.0) (2025-05-19)
+
+### Features
+
+- **key_manager:** add filtering and asymmetric algos ([#2125](https://github.com/scaleway/scaleway-sdk-js/issues/2125)) ([46d73be](https://github.com/scaleway/scaleway-sdk-js/commit/46d73be4a0378974e51db33994219d57d7f4bb70))
+
+## [1.1.1](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-key-manager@1.1.0...@scaleway/sdk-key-manager@1.1.1) (2025-05-13)
+
+**Note:** Version bump only for package @scaleway/sdk-key-manager
+
 # [1.1.0](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-key-manager@1.0.1...@scaleway/sdk-key-manager@1.1.0) (2025-05-13)
 
 ### Bug Fixes