+++ packages_generated/lb/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 1.2.0 (2025-05-26)
+
+### Features
+
+- **lb:** add access_logs field to frontend ([#2153](https://github.com/scaleway/scaleway-sdk-js/issues/2153)) ([332571e](https://github.com/scaleway/scaleway-sdk-js/commit/332571e85284dc79f5f0dd554891af1752e06a02))
+
+## [1.1.1](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-lb@1.1.0...@scaleway/sdk-lb@1.1.1) (2025-05-13)
+
+**Note:** Version bump only for package @scaleway/sdk-lb
+
 # [1.1.0](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-lb@1.0.1...@scaleway/sdk-lb@1.1.0) (2025-05-13)
 
 ### Bug Fixes