+++ packages_generated/mongodb/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.5](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-mongodb@1.0.4...@scaleway/sdk-mongodb@1.0.5) (2025-05-26)
+
+**Note:** Version bump only for package @scaleway/sdk-mongodb
+
+## [1.0.4](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-mongodb@1.0.3...@scaleway/sdk-mongodb@1.0.4) (2025-05-19)
+
+### Bug Fixes
+
+- **mongodb:** add the new SetUserRole endpoint on the documentation ([#2142](https://github.com/scaleway/scaleway-sdk-js/issues/2142)) ([c17a14a](https://github.com/scaleway/scaleway-sdk-js/commit/c17a14ac5c64e6e3c5132f6bf5b31c45c48cfa1b))
+
+## [1.0.3](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-mongodb@1.0.2...@scaleway/sdk-mongodb@1.0.3) (2025-05-13)
+
+**Note:** Version bump only for package @scaleway/sdk-mongodb
+
 ## [1.0.2](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-mongodb@1.0.1...@scaleway/sdk-mongodb@1.0.2) (2025-05-13)
 
 ### Bug Fixes