+++ packages_generated/vpc/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-vpc@1.1.0...@scaleway/sdk-vpc@1.1.1) (2025-05-26)
+
+**Note:** Version bump only for package @scaleway/sdk-vpc
+
+# [1.1.0](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-vpc@1.0.3...@scaleway/sdk-vpc@1.1.0) (2025-05-19)
+
+### Features
+
+- **vpc:** support defaultRoutePropagationEnabled ([#2140](https://github.com/scaleway/scaleway-sdk-js/issues/2140)) ([ab14ab7](https://github.com/scaleway/scaleway-sdk-js/commit/ab14ab71cfe333a30c9a908f5cf8ff70c0f202b2))
+
+## [1.0.3](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-vpc@1.0.2...@scaleway/sdk-vpc@1.0.3) (2025-05-13)
+
+**Note:** Version bump only for package @scaleway/sdk-vpc
+
 ## [1.0.2](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/sdk-vpc@1.0.1...@scaleway/sdk-vpc@1.0.2) (2025-05-13)
 
 ### Bug Fixes